Initialize OGL context
OpenGL is initialzied in this commit. Basic rendering with (very)
basic shaders is performed.

To acquire the context on Windows, regular HDC and HGLRC acquisition
for Window is performed.

To acquire the context on Linux, glX extension extraction is needed
to choose FrameBuffer configuration. These are then used in Window.

Verification:
    Build, run. Black window with colored triangle should be visible.

Change-Id: I8d45d09ad8f4c044cda7ad9147ef05e3b4623595