Update patch set 2
Patch Set 2:

(3 comments)

Ok, so I tweaked it a LOT:
 1. We have AnyException base class - we use it in catches
 2. We have Exception<ExceptionType> template class
 3. We have ExceptionType enum class that holds all types of exceptions
 4. We have ExceptionHelper structure for binding strings to template
 5. We have THROW macro, that lets us use LINE and FILE vars

Ad.1. AnyException is just a dummy class so we don't put Exception<ExceptionType::Math> in catch statement

Ad.2. We have Exception template class. Only difference is a type string - the string that appears at the begining of the exception message (e.x. for ExceptionType::Math it's "MathException").

Ad.3. Whenever we need new exception we add a new value to the ExceptionType enum class. Default type is "Exception", if we want another one, we need to specify ExceptionHelper struct for it.

Ad.4. ExceptionHelper is a helper structure that let's us bind string to template class. It works simple: for any ExceptionType::X we have ExceptionHelper<ExceptionType::X>::type that is a string used in Exception<ExceptionType::X>.

Ad.5. Simple macro that takes variadic number of args. First argument is ExceptionType::X, rest is formatted error message and its arguments.


Unfortunately due to it being a template, it MUST have declaration and definition in the same file. Afaik for templates it's acceptable. Too bad it's a whole class though - a lot of code :/

Patch-set: 2
Label: Code-Review=0