Update patch set 3
Patch Set 3: Code-Review+2

There happen to be A LOT of caves...how about we make liiiiike...~3 times less caves? Cause looking at it now makes me think "Dude...how does it NOT collapse?".

I verified it, but will leave it with 0, so as not to merge it accidentaly.

Patch-set: 3
Label: Code-Review=+2