Implement Chunk saving
When chunk is generated, it checks for saved file on the disk.
If not found, then chunk is generated in a standard way. Otherwise it's
loaded from the file.
Chunks are encoded using simple RLE algorithm.

Verification: Build and run. Check logs to see if chunks are generated
and saved or are they read from file. Check differences in speed.